Move dkjson.lua to lib directory so that script_manager doesn't try to add it as an executable.  Update git-annex.lua to look in the lib directory for dkjson.
